fixed constructor type

+++ include/Color.hpp
@@ -12,7 +12,7 @@ namespace Birb
 		Color(const int& r, const int& g, const int& b, const int& a);
 
 		Color(const float& r, const float& g, const float& b);
-		Color(const float& r, const float& g, const float& b, const int& a);
+		Color(const float& r, const float& g, const float& b, const float& a);
 
 		Color(const Vector3Int &rgb); ///< Uses values between 0 - 255
 		Color(const Vector3 &rgb);	  ///< Uses values between 0.0 - 1.0
